third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/access-control-basic-preflight-denied.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Tests async XHR preflight denial</title>
    <!--The original test addressed a more specific issue involving caching,
        but that issue has since been resolved.
        We maintain this test as a basic test of invalid preflight denial.
        Please refer to the comment in the following link for more information:
        https://chromium-review.googlesource.com/c/chromium/src/+/630338#message-0280542b95c9b0f82b121dc373320c04fcaece31
    -->
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script type="text/javascript">
    async_test((test) => {
      const xhr = new XMLHttpRequest;
      xhr.onerror = test.unreached_func("Unexpected error.");
      xhr.onreadystatechange = test.step_func(() => {
        if (xhr.readyState == XMLHttpRequest.OPENED) {
          assert_equals(xhr.status, 0);
        } else if (xhr.readyState == XMLHttpRequest.DONE) {
          test.done();
        }
      });

      xhr.open("FOO", get_host_info().HTTP_REMOTE_ORIGIN +
          "/XMLHttpRequest/resources/access-control-basic-denied.py");
      xhr.send();
    }, "Preflighted async XHR request");
    </script>
  </body>
</html>
