LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/access-control-basic-preflight-denied-expected.txt

CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/XMLHttpRequest/resources/access-control-basic-denied.py due to access control checks.

PASS Tests async XHR preflight denial due to lack of CORS headers 
